Removed ingredient parsing
The attempt to parse a string with ingredients was probably misguided;
experience with manual entry of printed recipes has shown that this is
really not the kind of thing that can be automated. The current admin
interface for entering ingredients is pretty good, and actually requires
less typing (and less thinking) than plain-text entry of
specially-formatted ingredients. Maybe it'll be useful some day, but for
now it should go away.