Overhauled NutritionInfo
Refactored NutritionInfo into an inheritance hierarchy, with the basic
NutritionInfo model having only the grams/milligrams of nutrients, and
new model FoodNutritionInfo taking care of food/quantity/unit.

Also introduced new RecipeNutritionInfo and IngredientNutritionInfo
derivatives, with hooks to recalculate them when the respective model
instance is saved. This allows persistent storage of nutrition
information for each ingredient and recipe, preventing inefficient
recalculation of them all the time.

It also allows a new feature that I've been meaning to implement (now
implemented), whereby hovering the mouse over an ingredient in a recipe
shows a popup with its nutrition information.

Unit tests were adjusted accordingly. Some new tests will need to be
written to cover the save() and recalculate() hooks, but for the most
part it seems to be working as intended now.